requirements.txt
Flask
python-dotenv
openai
APScheduler
requests